refactor aide to its own directory
